packages/agents/bidi-agent/pyproject.toml
[project]
name = "idp_v2.bidi-agent"
version = "0.1.0"
description = "Bidi Agent"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "asyncio>=4.0.0",
    "strands-agents[bidi-all]>=1.25.0",
    "fastapi[standard]>=0.123.0",
    "uvicorn>=0.34.0",
    "pydantic-settings>=2.0.0",
    "boto3>=1.35.0",
    "pytz>=2024.1",
    "httpx>=0.27.0",
    "mcp>=1.22.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
exclude = [".ruff_cache", ".svn", ".tox", ".venv", "dist"]
line-length = 120
indent-width = 4

  [tool.ruff.lint]
  select = ["E", "F", "UP", "B", "SIM", "I"]
  ignore = []
  fixable = ["ALL"]
  unfixable = []
